third_party/WebKit/LayoutTests/http/tests/accessibility/slow-document-load-expected.txt

This tests that accessibility events will be fired after the first layout, rather than only after the document loads.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS elapsedTime is >= 900
PASS gotLayoutCompleteBeforeOneSecond is true
PASS successfullyParsed is true

TEST COMPLETE
